+++ tomcat/src/main/java/com/techcourse/controller/RootRequestController.java
@@ -1,12 +1,13 @@
-package com.techcourse.handler;
+package com.techcourse.controller;
 
 import com.techcourse.exception.UncheckedServletException;
 import java.nio.charset.StandardCharsets;
 import org.apache.coyote.HttpRequest;
 import org.apache.coyote.HttpResponse;
+import org.apache.coyote.http11.AbstractRequestHandler;
 import org.apache.coyote.http11.MimeType;
 
-public class RootRequestHandler extends AbstractRequestHandler {
+public class RootRequestController extends AbstractRequestHandler {
 
     private static final String ROOT_RESOURCE = "Hello world!";
