+++ tomcat/src/test/java/com/techcourse/controller/LoginRequestControllerTest.java
@@ -1,9 +1,8 @@
-package com.techcourse.handler;
+package com.techcourse.controller;
+
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.techcourse.exception.UncheckedServletException;
 import java.util.List;
 import org.apache.coyote.HttpRequest;
 import org.apache.coyote.HttpResponse;
@@ -16,18 +15,36 @@ import org.apache.coyote.http11.response.Http11Response;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-public class SignupRequestHandlerTest {
+class LoginRequestControllerTest {
 
-    private RequestHandler handler = new SignupRequestHandler();
+    private RequestHandler handler = new LoginRequestController();
 
-    @DisplayName("회원가입을 하면 /index.html로 리다이렉트 한다.")
+    @DisplayName("로그인을 하면 /index.html로 리다이렉트 한다.")
     @Test
     void handle1() throws Exception {
-        String body = "account=test&email=test@test&password=test";
-        Http11RequestLine requestLine = new Http11RequestLine("POST /register HTTP/1.1");
+        String body = "account=gugu&password=password";
+        Http11RequestLine requestLine = new Http11RequestLine("POST /login HTTP/1.1");
         Http11RequestHeaders headers = new Http11RequestHeaders(
                 List.of("Content-Length: " + body.getBytes().length));
         Http11RequestBody requestBody = new Http11RequestBody(body);
+        HttpRequest request = new Http11Request(requestLine, headers, requestBody);
+        HttpResponse response = new Http11Response();
+        handler.handle(request, response);
+        String actual = response.getResponseMessage();
+        String expectedResponseLine = "HTTP/1.1 302 Found";
+        String expectedLocation = "/index.html";
+        assertThat(actual).contains(expectedResponseLine);
+        assertThat(actual).contains(expectedLocation);
+    }
+
+    @DisplayName("쿠키에 세션 id가 존재하면, /login GET 요청 시 /index.html로 리다이렉트 된다.")
+    @Test
+    void handle2() throws Exception {
+        Http11RequestLine requestLine = new Http11RequestLine("GET /login HTTP/1.1");
+        Http11RequestHeaders headers = new Http11RequestHeaders(
+                List.of("Cookie: JSESSIONID=123")
+        );
+        Http11RequestBody requestBody = new Http11RequestBody("");
         HttpRequest request = new Http11Request(requestLine, headers, requestBody);
         HttpResponse response = new Http11Response();
 
@@ -36,25 +53,8 @@ public class SignupRequestHandlerTest {
         String actual = response.getResponseMessage();
         String expectedResponseLine = "HTTP/1.1 302 Found";
         String expectedLocation = "/index.html";
+
         assertThat(actual).contains(expectedResponseLine);
         assertThat(actual).contains(expectedLocation);
-    }
-
-    @DisplayName("회원가입에 실패하면, 예외가 발생한다.")
-    @Test
-    void handle2() {
-        //구구 계정 이미 있음
-        String body = "account=gugu&email=test@test&password=password";
-        Http11RequestLine requestLine = new Http11RequestLine("POST /register HTTP/1.1");
-        Http11RequestHeaders headers = new Http11RequestHeaders(
-                List.of("Content-Length: " + body.getBytes().length));
-        Http11RequestBody requestBody = new Http11RequestBody(body);
-        HttpRequest request = new Http11Request(requestLine, headers, requestBody);
-        HttpResponse response = new Http11Response();
-
-        assertThatThrownBy(() -> handler.handle(request, response))
-                .isInstanceOf(UncheckedServletException.class)
-                .hasMessage("java.lang.IllegalStateException: 이미 존재하는 아이디 입니다.");
-
     }
 }
